ch08-e-shiritori/pyproject.toml
[project]
name = "e-shiritori"
version = "0.1.0"
description = "Chainlit + Gemini API で作る絵しりとり Web アプリ"
requires-python = ">=3.13"
dependencies = [
    "chainlit==2.9.6",
    "google-genai>=1.65.0",
    "python-dotenv>=1.2.1",
]
